[AS-1703] feat: add durationInBlock to solve arbitrum block time < 350ms
