+++ hardhat.config.js
@@ -6,6 +6,8 @@ require('solidity-coverage');
 require('@openzeppelin/hardhat-upgrades');
 require('hardhat-contract-sizer');
 require('hardhat-test-utils');
+require('hardhat-storage-layout');
+
 require('./tasks/interface');
 require('./tasks/export');
 require('./tasks/decodeUniswap');
