+++ package.json
@@ -56,6 +56,7 @@
     "hardhat": "^2.11.2",
     "hardhat-abi-exporter": "^2.8.0",
     "hardhat-contract-sizer": "^2.4.0",
+    "hardhat-storage-layout": "^0.1.7",
     "hardhat-test-utils": "^0.1.1",
     "lodash": "^4.17.21",
     "minimatch": "^5.0.1"
