+++ test/unit/GovernFactory.js
@@ -65,14 +65,14 @@ describe('GovernFactory.sol - test/unit/GovernFactory.js', function () {
 
   describe('createGovern()', function () {
     it('creates a new govern', async function () {
-      const tx = await governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE');
+      const tx = await governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE', 0);
       const { govern: governAddress } = await findEventArgs(tx, 'CreateGovern');
       expect(await governFactory.governMap(creator.address, 'mockName')).to.equal(governAddress);
     });
 
     it('throws "error" error if name is duplicated', async function () {
-      await governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE');
-      await expect(governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE')).to.be.revertedWith('error');
+      await governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE', 5);
+      await expect(governFactory.createGovern('mockName', 0, 0, 0, '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE', 5)).to.be.revertedWith('error');
     });
   });
 });
